fix(backend): resolve broken import paths after service refactoring
